requirements.txt
jupyter==1.0.0
matplotlib==3.1.1
numpy==1.17.0
Pillow==6.2.0
sklearn==0.0
snorkel==0.9.3
networkx>=2.2,<2.4

